Rename option reducer file
